additional changes included

+++ src/main/java/com/shoppingcart/app/entity/Product.java
@@ -1,15 +1,12 @@
 package com.shoppingcart.app.entity;
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigDecimal;
 
 @Entity(name = "PRODUCT")
 @Data
@@ -27,5 +24,5 @@ public class Product {
     private String description;
 
     @Column(name = "PRODUCT_COST", length = 30)
-    private BigDecimal cost;
+    private Double cost;
 }
